Готовимся к рисункам схем включения

+++ sources/PLoader/src/FlashDrive/FlashDrive.cpp
@@ -416,7 +416,7 @@ void Upgrade()
         size -= readedBytes;
         address += static_cast<uint>(readedBytes);
 
-        percentsUpdate = 1.0F - static_cast<float>(size) / fullSize;
+        percentsUpdate = 1.0F - static_cast<float>(size) / (float)fullSize;
     }
 
     FDrive::CloseOpenedFile();
